refactor(spring): extract OpenAI provider auto-config (Epic 6.3, step 2b)
New module tramai-spring-provider-openai owns OpenAI and OpenAI-compatible
construction from tramai.providers.openai.* / openai-compatible.* properties.
It contributes SpringConfiguredModelProvider beans; core TramaiAutoConfiguration
collects them as property-generated providers via ObjectProvider and no longer
imports any OpenAI class. @AutoConfigureBefore(TramaiAutoConfiguration) +
@ConditionalOnMissingBean(Tramai) preserves the original ordering (providers
visible at assembly) and the user-Tramai back-off.

- OpenAiProviderAutoConfiguration + own AutoConfiguration.imports
- core TramaiAutoConfiguration: openai construction removed; propertyProviders
  now = inline (anthropic, ollama) + SpringConfiguredModelProvider beans
- characterization tests moved to tramai-spring test source set (integration
  home): every runner registers OpenAiProviderAutoConfiguration — assertion
  content unchanged, harness wiring only
- settings, module-catalog, publishable list, BOM updated; api dumps

61/61 tests green; verifyPr green.
