+++ tramai-spring-provider-openai/build.gradle.kts
@@ -19,20 +19,18 @@ kotlin {
     }
 }
 
-// Epic 6.3 compatibility facade: generic Spring integration moved to
-// tramai-spring-core. This artifact deliberately no longer pulls provider
-// adapters or secret-backend SDKs; consumers select them explicitly.
-//
-// Its test source set is the integration-test home: the characterization
-// suite runs against core + the adapter modules under test.
+// OpenAI / OpenAI-compatible provider adapter for Spring (Epic 6.3).
+// Owns OpenAI construction only; generic assembly lives in tramai-spring-core.
 dependencies {
     api(project(":tramai-spring-core"))
+    implementation(project(":tramai-openai"))
 
-    testImplementation(project(":tramai-spring-provider-openai"))
+    implementation(libs.spring.context)
+    implementation(libs.spring.boot.autoconfigure)
+
     testImplementation(libs.coroutines.core)
     testImplementation(libs.kotlin.test.junit5)
     testImplementation(libs.spring.boot.starter.test)
-    testImplementation(project(":tramai-security"))
 }
 
 tasks.test {
